fix post action
the POST endpoints only returns the subject UUID and not a complete, nested subject object representation.
Untl the API is fixed, we make a subsequent GET call

also, if the modification includes a subject, we pass only the UUID to $axios, as the endpoint requires `subject` to be a UUID and not an object.